fixed vector2 variable assigning; began adding mexicantraininfo; changed exception handling; added message to input.choice

+++ src/Options/Games/MexicanTrain/Domino.cs
@@ -6,15 +6,15 @@ namespace B.Options.Games.MexicanTrain
         public int Value2 => GetValue(_val2);
         public int Total => Value1 + Value2;
 
-        private readonly byte _val1;
-        private readonly byte _val2;
+        private readonly int _val1;
+        private readonly int _val2;
 
-        public Domino(byte val1, byte val2)
+        public Domino(int val1, int val2)
         {
             _val1 = val1;
             _val2 = val2;
         }
 
-        private int GetValue(byte val) => val == 0 ? 25 : val;
+        private int GetValue(int val) => val == 0 ? 25 : val;
     }
 }
